1) Make GravitationalAcceleration Type-stable 2) Ensure that MatParams is isbits, which is crucial to send the struct to the GPU.
Doing this required removing the equations from the struct, as they are strings of a form that makes it not isbits. We will have to reorganise the code accordingly at a later stage to allow reconstructing comments/equations/references from the definitions in the MatParam struct. As such equations are not used in actual computations it makes sense to keep them separate anyways.
Also note that it required saving the name of a field as a NTuple of Char to make it isbittype.