correct first try at table
First commit with "first try" commit message was the wrong code. This is the state the table code is at right now. All parameter values are plotted the first phase column which still needs to be fixed. Otherwise it should output a latex ".tex" file in the same directory as the file is.